+++ src/components/ImageCard/index.tsx
@@ -1,6 +1,6 @@
-import { ImageDataType } from '~apis/images';
-import * as styles from './ImageCard.module.css';
 import { memo } from 'react';
+import { ImageDataType } from '~lib/apis/images';
+import * as styles from './ImageCard.module.css';
 
 interface ImageCardProps {
   data: ImageDataType;
@@ -13,6 +13,7 @@ export default memo(function ImageCard({ data, isAnimating }: ImageCardProps) {
       className={`${styles.imageContainer} ${isAnimating ? styles.upAnimation : ''}`}
     >
       <img
+        loading="lazy"
         src={data.urls.small_s3}
         alt={data.alt_description}
         className={styles.image}
