+++ src/lib/mocks/browser.ts
